Add audio indicator and button to close, pin tabs
Fixes #50
Fixes #53
